Fix (Math): Fix incorrect type conversion in SignedColor-related functions

+++ include/raynder/math_helpers.h
@@ -24,25 +24,25 @@ namespace Math {
 
     inline SignedColor subtract_color(const SignedColor A, const SignedColor B) {
         return SignedColor{
-            (int8_t)(A.r - B.r),
-            (int8_t)(B.g - B.g),
-            (int8_t)(A.b - B.b)
+            (int16_t)(A.r - B.r),
+            (int16_t)(B.g - B.g),
+            (int16_t)(A.b - B.b)
         };
     }
 
     inline SignedColor add_color(const SignedColor A, const SignedColor B) {
         return SignedColor{
-            (int8_t)(A.r + B.r),
-            (int8_t)(A.g + B.g),
-            (int8_t)(A.b + B.b)
+            (int16_t)(A.r + B.r),
+            (int16_t)(A.g + B.g),
+            (int16_t)(A.b + B.b)
         };
     }
 
     inline SignedColor multiply_color(const SignedColor A, const float c) {
         return SignedColor{
-            (int8_t)(A.r * c),
-            (int8_t)(A.g * c),
-            (int8_t)(A.b * c)
+            (int16_t)(A.r * c),
+            (int16_t)(A.g * c),
+            (int16_t)(A.b * c)
         };
     }
